feat: clean up image preprocessing
